Restructured files, + index.ts for quick imports. Basic responsiveness
