+++ frontend/foodieland-app/app/(auth)/page.module.css
@@ -22,7 +22,7 @@
     height: 100%;
     display: flex;
     align-items: center;
-    justify-content: end;
+    justify-content: flex-end;
 }
 .loginRightContent {
     display: flex;
@@ -53,7 +53,7 @@
     display: flex;
     flex-direction: column;
     gap: 20px;
-    align-items: start;
+    align-items: flex-start;
 }
 
 .loginContent > div > h1 {
@@ -76,4 +76,10 @@
 .loginFooter > p {
     color: rgba(0, 0, 0, 0.7);
     letter-spacing: 0.3px;
+}
+
+@media (max-width: 900px) {
+    .loginLeft {
+        flex: 0 1 100%;
+    }
 }
