+++ frontend/foodieland-app/components/RegisterForm.tsx
@@ -1,8 +1,17 @@
-import styles from '../form.module.css'
-export default function LoginForm()
+import styles from './form.module.css'
+export default function RegisterForm()
 {
     return (
         <form>
+            <div className={styles.inputGroup}>
+                <label htmlFor="fullName">Full name</label>
+                <input
+                    type="text"
+                    id="fullName"
+                    placeholder="John Doe"
+                    required
+                />
+            </div>
             <div className={styles.inputGroup}>
                 <label htmlFor="email">E-mail</label>
                 <input
@@ -21,7 +30,7 @@ export default function LoginForm()
                     required
                 />
             </div>
-            <button type="submit" className={`${styles.submitButton} btnMain`}>Log In</button>
+            <button type="submit" className={`${styles.submitButton} btnMain`}>Sign Up</button>
         </form>
     );
 }
