+++ frontend/foodieland-app/components/index.ts
@@ -1,0 +1,13 @@
+import Logo from "@/components/Logo";
+import Button from "@/components/Button";
+import LoginForm from "@/components/LoginForm";
+import RegisterForm from "@/components/RegisterForm";
+import Promosection from "@/components/Promosection";
+
+export {
+  Logo,
+  Button,
+  LoginForm,
+  RegisterForm,
+  Promosection
+};
